issue #11, render target supports buffers control interface

+++ lite3d/include/lite3d/lite3d_vbo.h
@@ -44,13 +44,13 @@ typedef struct lite3d_vbo
 typedef struct lite3d_vao
 {
     uint32_t vaoID;
-    size_t elementsCount;
-    size_t indexesCount;
-    size_t indexesSize;
-    size_t indexesOffset;
-    size_t verticesCount;
-    size_t verticesSize;
-    size_t verticesOffset;
+    uint32_t elementsCount;
+    uint32_t indexesCount;
+    uint32_t indexesSize;
+    uint32_t indexesOffset;
+    uint32_t verticesCount;
+    uint32_t verticesSize;
+    uint32_t verticesOffset;
     uint16_t elementType; /* GL_POINTS, GL_LINES, GL_TRIANGLES (GL value) */
     uint16_t indexType; /* Byte, short, int (GL value) */
 } lite3d_vao;
